+++ go.mod
@@ -1,3 +1,3 @@
-module bitcoin-rate-app
+module bitcoinrateapp
 
 go 1.20
